LinearSearchAlgorithm
This Java program implements a simple linear search algorithm to find the number of occurrences of a given value in an array. The linearSearch class contains an array (array) and the number of elements in the array (nElems). The constructor initializes these values. The lSearch method iterates through the array elements and counts the occurrences of the given value. The main method creates an instance of linearSearch with an array {10,20,30,50,40,50,90,80} and its length. It then takes user input to find the number of occurrences of a value in the array using the lSearch method. If the value is not found, it prints a message indicating that the value is not in the array; otherwise, it prints the number of times the value was found.